requirements.txt
numpy>=1.21
scipy>=1.7
pyyaml>=6.0
